test: move contract tests into integration tests
