+++ crates/pop-contracts/src/call.rs
@@ -267,260 +267,3 @@ pub fn get_call_payload(
 	.encode_call_data_to(&call_exec.client().metadata(), &mut encoded_data)?;
 	Ok(encoded_data)
 }
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use crate::{
-		contracts_node_generator, dry_run_gas_estimate_instantiate, errors::Error,
-		instantiate_smart_contract, mock_build_process, new_environment, run_contracts_node,
-		set_up_deployment, Bytes, UpOpts,
-	};
-	use anyhow::Result;
-	use pop_common::{find_free_port, set_executable_permission};
-	use std::{env, process::Command, time::Duration};
-	use tokio::time::sleep;
-
-	#[cfg(feature = "v5")]
-	const CONTRACT_ADDRESS: &str = "5CLPm1CeUvJhZ8GCDZCR7nWZ2m3XXe4X5MtAQK69zEjut36A";
-	#[cfg(feature = "v5")]
-	const CONTRACTS_NETWORK_URL: &str = "wss://rpc1.paseo.popnetwork.xyz";
-	#[cfg(feature = "v6")]
-	const CONTRACT_ADDRESS: &str = "0x4f04054746fb19d3b027f5fe1ca5e87a68b49bac";
-	#[cfg(feature = "v6")]
-	const CONTRACTS_NETWORK_URL: &str = "wss://westend-asset-hub-rpc.polkadot.io";
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn test_set_up_call() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join("./tests/files/testing.contract"),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-
-		let call_opts = CallOpts {
-			path: Some(temp_dir.path().join("testing")),
-			contract: CONTRACT_ADDRESS.to_string(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		};
-		let call = set_up_call(call_opts).await?;
-		assert_eq!(call.message(), "get");
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn test_set_up_call_from_artifact_file() -> Result<()> {
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		let call_opts = CallOpts {
-			path: Some(current_dir.join("./tests/files/testing.json")),
-			contract: CONTRACT_ADDRESS.to_string(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		};
-		let call = set_up_call(call_opts).await?;
-		assert_eq!(call.message(), "get");
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn test_set_up_call_error_contract_not_build() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let call_opts = CallOpts {
-			path: Some(temp_dir.path().join("testing")),
-			contract: CONTRACT_ADDRESS.to_string(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		};
-		assert!(
-			matches!(set_up_call(call_opts).await, Err(Error::AnyhowError(message)) if message.root_cause().to_string() == "Failed to find any contract artifacts in target directory. \nRun `cargo contract build --release` to generate the artifacts.")
-		);
-		Ok(())
-	}
-	// TODO: Integration test
-	#[tokio::test]
-	async fn test_set_up_call_fails_no_smart_contract_directory() -> Result<()> {
-		let call_opts = CallOpts {
-			path: None,
-			contract: CONTRACT_ADDRESS.to_string(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		};
-		assert!(
-			matches!(set_up_call(call_opts).await, Err(Error::AnyhowError(message)) if message.root_cause().to_string() == "No 'ink' dependency found")
-		);
-		Ok(())
-	}
-	// TODO: Integration test
-	#[tokio::test]
-	async fn test_dry_run_call_error_contract_not_deployed() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join("./tests/files/testing.contract"),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-
-		let call_opts = CallOpts {
-			path: Some(temp_dir.path().join("testing")),
-			contract: CONTRACT_ADDRESS.to_string(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		};
-		let call = set_up_call(call_opts).await?;
-		assert!(matches!(dry_run_call(&call).await, Err(Error::DryRunCallContractError(..))));
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn test_dry_run_estimate_call_error_contract_not_deployed() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join("./tests/files/testing.contract"),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-
-		let call_opts = CallOpts {
-			path: Some(temp_dir.path().join("testing")),
-			contract: CONTRACT_ADDRESS.to_string(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		};
-		let call = set_up_call(call_opts).await?;
-		assert!(matches!(
-			dry_run_gas_estimate_call(&call).await,
-			Err(Error::DryRunCallContractError(..))
-		));
-		Ok(())
-	}
-	// TODO: Integration test, running local node
-	#[tokio::test]
-	async fn call_works() -> Result<()> {
-		let random_port = find_free_port(None);
-		let localhost_url = format!("ws://127.0.0.1:{}", random_port);
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		#[cfg(feature = "v5")]
-		let contract_file = "./tests/files/testing_wasm.contract";
-		#[cfg(feature = "v6")]
-		let contract_file = "./tests/files/testing.contract";
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(contract_file),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-
-		let cache = temp_dir.path().join("");
-
-		let binary = contracts_node_generator(cache.clone(), None).await?;
-		binary.source(false, &(), true).await?;
-		set_executable_permission(binary.path())?;
-		let process = run_contracts_node(binary.path(), None, random_port).await?;
-		// Wait 5 secs more to give time for the node to be ready
-		sleep(Duration::from_millis(5000)).await;
-		// Instantiate a Smart Contract.
-		let instantiate_exec = set_up_deployment(UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "0".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: Some(Bytes::from(vec![0x00])),
-			url: Url::parse(&localhost_url)?,
-			suri: "//Alice".to_string(),
-		})
-		.await?;
-		let weight = dry_run_gas_estimate_instantiate(&instantiate_exec).await?;
-		let contract_info = instantiate_smart_contract(instantiate_exec, weight).await?;
-		// Test querying a value.
-		let query_exec = set_up_call(CallOpts {
-			path: Some(temp_dir.path().join("testing")),
-			contract: contract_info.address.clone(),
-			message: "get".to_string(),
-			args: [].to_vec(),
-			value: "0".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(&localhost_url)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		})
-		.await?;
-		let mut query = dry_run_call(&query_exec).await?;
-		assert_eq!(query, "Ok(false)");
-		// Test extrinsic execution by flipping the value.
-		let call_exec = set_up_call(CallOpts {
-			path: Some(temp_dir.path().join("testing")),
-			contract: contract_info.address,
-			message: "flip".to_string(),
-			args: [].to_vec(),
-			value: "0".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			url: Url::parse(&localhost_url)?,
-			suri: "//Alice".to_string(),
-			execute: false,
-		})
-		.await?;
-		let weight = dry_run_gas_estimate_call(&call_exec).await?;
-		assert!(weight.ref_time() > 0);
-		assert!(weight.proof_size() > 0);
-		call_smart_contract(call_exec, weight, &Url::parse(&localhost_url)?).await?;
-		// Assert that the value has been flipped.
-		query = dry_run_call(&query_exec).await?;
-		assert_eq!(query, "Ok(true)");
-		// Stop the process contracts-node
-		Command::new("kill")
-			.args(["-s", "TERM", &process.id().to_string()])
-			.spawn()?
-			.wait()?;
-
-		Ok(())
-	}
-}
