+++ crates/pop-contracts/src/node.rs
@@ -232,8 +232,6 @@ fn release_directory_by_target(binary: &str) -> Result<Vec<ArchiveFileSpec>, Err
 mod tests {
 	use super::*;
 	use anyhow::{Error, Result};
-	use pop_common::find_free_port;
-	use std::process::Command;
 
 	#[tokio::test]
 	async fn directory_path_by_target() -> Result<()> {
@@ -248,7 +246,6 @@ mod tests {
 		Ok(())
 	}
 
-	// TODO: Integration test
 	#[tokio::test]
 	async fn is_chain_alive_works() -> Result<(), Error> {
 		let local_url = url::Url::parse("ws://wrong")?;
@@ -294,40 +291,6 @@ mod tests {
 				cache == cache
 			));
 		}
-		Ok(())
-	}
-
-	// TODO: Integration test, slow (30s)
-	#[tokio::test]
-	async fn run_contracts_node_works() -> Result<(), Error> {
-		let random_port = find_free_port(None);
-		let localhost_url = format!("ws://127.0.0.1:{}", random_port);
-		let local_url = url::Url::parse(&localhost_url)?;
-
-		let temp_dir = tempfile::tempdir().expect("Could not create temp dir");
-		let cache = temp_dir.path().join("");
-
-		#[cfg(feature = "v5")]
-		let version = "v0.42.0";
-		#[cfg(feature = "v6")]
-		let version = "v0.43.0";
-		let binary = contracts_node_generator(cache.clone(), Some(version)).await?;
-		binary.source(false, &(), true).await?;
-		let process = run_contracts_node(binary.path(), None, random_port).await?;
-
-		// Check if the node is alive
-		assert!(is_chain_alive(local_url).await?);
-		#[cfg(feature = "v5")]
-		assert!(cache.join("substrate-contracts-node-v0.42.0").exists());
-		#[cfg(feature = "v6")]
-		assert!(cache.join("ink-node-v0.43.0").exists());
-		assert!(!cache.join("artifacts").exists());
-		// Stop the process contracts-node
-		Command::new("kill")
-			.args(["-s", "TERM", &process.id().to_string()])
-			.spawn()?
-			.wait()?;
-
 		Ok(())
 	}
 }
