+++ crates/pop-contracts/src/up.rs
@@ -563,291 +563,3 @@ pub fn get_code_hash_from_event<C: Config>(
 		)))
 	}
 }
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	#[cfg(feature = "v5")]
-	use crate::errors::Error;
-	use crate::{
-		contracts_node_generator, mock_build_process, new_environment, run_contracts_node,
-	};
-	use anyhow::Result;
-	use pop_common::{find_free_port, set_executable_permission};
-	use std::{env, process::Command, time::Duration};
-	use subxt::{
-		config::{substrate::BlakeTwo256, Hasher},
-		utils::H256,
-	};
-	use tokio::time::sleep;
-	use url::Url;
-
-	#[cfg(feature = "v5")]
-	const CONTRACT_FILE: &str = "./tests/files/testing_wasm.contract";
-	#[cfg(feature = "v6")]
-	const CONTRACT_FILE: &str = "./tests/files/testing.contract";
-	#[cfg(feature = "v5")]
-	const CONTRACTS_NETWORK_URL: &str = "wss://rpc1.paseo.popnetwork.xyz";
-	#[cfg(feature = "v6")]
-	const CONTRACTS_NETWORK_URL: &str = "wss://westend-asset-hub-rpc.polkadot.io";
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn set_up_deployment_works() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-		let up_opts = UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-		};
-		set_up_deployment(up_opts).await?;
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn set_up_upload_works() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-		let up_opts = UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-		};
-		set_up_upload(up_opts).await?;
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn get_payload_works() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-		let up_opts = UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-		};
-		let contract_code = get_contract_code(up_opts.path.as_ref())?;
-		#[cfg(feature = "v5")]
-		let call_data = get_upload_payload(contract_code, CONTRACTS_NETWORK_URL).await?;
-		#[cfg(feature = "v6")]
-		let call_data = {
-			let upload_exec = set_up_upload(up_opts).await?;
-			get_upload_payload(upload_exec, contract_code, CONTRACTS_NETWORK_URL).await?
-		};
-		let payload_hash = BlakeTwo256::hash(&call_data);
-		// We know that for the above opts the payload hash should be:
-		// 0x98c24584107b3a01d12e8e02c0bb634d15dc86123c44d186206813ede42f478d
-		#[cfg(feature = "v5")]
-		let hex_bytes = from_hex("98c24584107b3a01d12e8e02c0bb634d15dc86123c44d186206813ede42f478d")
-			.expect("Invalid hex string");
-		#[cfg(feature = "v6")]
-		let hex_bytes = from_hex("1e971a1ba0f3fe41c9c162ab30bb0ab9300108ddf32c4c4cdd01adf01638a76f")
-			.expect("Invalid hex string");
-
-		let hex_array: [u8; 32] = hex_bytes.try_into().expect("Expected 32-byte array");
-
-		// Create `H256` from the `[u8; 32]` array
-		let expected_hash = H256::from(hex_array);
-		assert_eq!(expected_hash, payload_hash);
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn dry_run_gas_estimate_instantiate_works() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-		let up_opts = UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "0".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-		};
-		let instantiate_exec = set_up_deployment(up_opts).await?;
-		let weight = dry_run_gas_estimate_instantiate(&instantiate_exec).await?;
-		assert!(weight.ref_time() > 0);
-		assert!(weight.proof_size() > 0);
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	#[cfg(feature = "v5")]
-	async fn dry_run_gas_estimate_instantiate_throw_custom_error() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-		let up_opts = UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "10000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-		};
-		let instantiate_exec = set_up_deployment(up_opts).await?;
-		assert!(matches!(
-			dry_run_gas_estimate_instantiate(&instantiate_exec).await,
-			Err(Error::DryRunUploadContractError(..))
-		));
-		Ok(())
-	}
-
-	// TODO: Integration test
-	#[tokio::test]
-	async fn dry_run_upload_throw_custom_error() -> Result<()> {
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-		let up_opts = UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(CONTRACTS_NETWORK_URL)?,
-			suri: "//Alice".to_string(),
-		};
-		let upload_exec = set_up_upload(up_opts).await?;
-		let upload_result = dry_run_upload(&upload_exec).await?;
-		assert!(!upload_result.code_hash.starts_with("0x0x"));
-		assert!(upload_result.code_hash.starts_with("0x"));
-		Ok(())
-	}
-
-	// TODO: Integration test, running local node
-	#[tokio::test]
-	async fn instantiate_and_upload() -> Result<()> {
-		let random_port = find_free_port(None);
-		let localhost_url = format!("ws://127.0.0.1:{}", random_port);
-		let temp_dir = new_environment("testing")?;
-		let current_dir = env::current_dir().expect("Failed to get current directory");
-		mock_build_process(
-			temp_dir.path().join("testing"),
-			current_dir.join(CONTRACT_FILE),
-			current_dir.join("./tests/files/testing.json"),
-		)?;
-
-		let cache = temp_dir.path().join("");
-
-		let binary = contracts_node_generator(cache.clone(), None).await?;
-		binary.source(false, &(), true).await?;
-		set_executable_permission(binary.path())?;
-		let process = run_contracts_node(binary.path(), None, random_port).await?;
-		// Wait 5 secs more to give time for the node to be ready
-		sleep(Duration::from_millis(5000)).await;
-		let upload_exec = set_up_upload(UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: [].to_vec(),
-			value: "1000".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: None,
-			url: Url::parse(&localhost_url)?,
-			suri: "//Alice".to_string(),
-		})
-		.await?;
-		// Only upload a Smart Contract
-		let upload_result = upload_smart_contract(&upload_exec).await?;
-		assert!(!upload_result.starts_with("0x0x"));
-		assert!(upload_result.starts_with("0x"));
-		// Error when Smart Contract has been already uploaded, only for ink!v5.
-		#[cfg(feature = "v5")]
-		assert!(matches!(
-			upload_smart_contract(&upload_exec).await,
-			Err(Error::UploadContractError(..))
-		));
-
-		// Instantiate a Smart Contract
-		let instantiate_exec = set_up_deployment(UpOpts {
-			path: Some(temp_dir.path().join("testing")),
-			constructor: "new".to_string(),
-			args: ["false".to_string()].to_vec(),
-			value: "0".to_string(),
-			gas_limit: None,
-			proof_size: None,
-			salt: Some(Bytes::from(vec![0x00])),
-			url: Url::parse(&localhost_url)?,
-			suri: "//Alice".to_string(),
-		})
-		.await?;
-		// First gas estimation
-		let weight = dry_run_gas_estimate_instantiate(&instantiate_exec).await?;
-		assert!(weight.ref_time() > 0);
-		assert!(weight.proof_size() > 0);
-		// Instantiate smart contract
-		let contract_info = instantiate_smart_contract(instantiate_exec, weight).await?;
-		#[cfg(feature = "v6")]
-		assert!(contract_info.address.starts_with("0x"));
-		#[cfg(feature = "v5")]
-		assert!(contract_info.address.starts_with("5"));
-		assert!(contract_info.code_hash.is_none());
-		// Stop the process contracts-node
-		Command::new("kill")
-			.args(["-s", "TERM", &process.id().to_string()])
-			.spawn()?
-			.wait()?;
-
-		Ok(())
-	}
-}
